refactor(core): import drain_background_tasks from its home module
The scheduler machinery moved to basic_memory.index.local_schedulers,
but deps/services.py kept a re-export alias so api/app.py,
mcp/server.py, and cli/commands/command_utils.py could keep importing
the moved runtime behavior from the composition root — contradicting
that module's own 'no runtime behavior of its own' docstring. Point the
three lifespan call sites at index.local_schedulers directly and drop
the alias; the CLI cleanup test now monkeypatches the owning module.
