+++ src/basic_memory/api/app.py
@@ -29,7 +29,7 @@ from basic_memory.api.v2.routers.project_router import (
 import logfire
 from basic_memory.cloud.note_content_materialization import drain_pending_materializations
 from basic_memory.config import init_api_logging
-from basic_memory.deps.services import drain_background_tasks
+from basic_memory.index.local_schedulers import drain_background_tasks
 from basic_memory.services.exceptions import EntityAlreadyExistsError
 from basic_memory.services.initialization import initialize_app
 from basic_memory.workspace_context import (
