+++ src/basic_memory/mcp/server.py
@@ -14,7 +14,7 @@ from basic_memory import db
 from basic_memory.cli.auth import CLIAuth
 from basic_memory.cloud.note_content_materialization import drain_pending_materializations
 from basic_memory.db import scoped_session
-from basic_memory.deps.services import drain_background_tasks
+from basic_memory.index.local_schedulers import drain_background_tasks
 from basic_memory.mcp.client_info import MCPClientInfoMiddleware
 from basic_memory.mcp.container import McpContainer, set_container
 from basic_memory.services.initialization import initialize_app
